cdk/alert_agent_stack.py
"""
CDK Stack for Real-Time Trafficking Alert Agent
Infrastructure will be implemented in task 7
"""

from aws_cdk import (
    Stack,
    aws_lambda as lambda_,
    aws_events as events,
    aws_events_targets as targets,
    aws_dynamodb as dynamodb,
    aws_iam as iam,
    Duration
)
from constructs import Construct


class AlertAgentStack(Stack):
    def __init__(self, scope: Construct, id: str, **kwargs) -> None:
        super().__init__(scope, id, **kwargs)
        
        # Infrastructure resources will be defined in task 7
        # Placeholder for:
        # - DynamoDB table
        # - Lambda function
        # - EventBridge rule
        # - IAM permissions
        pass
