requirements.txt
strands-agents>=0.1.0
boto3>=1.34.0
aws-cdk-lib>=2.100.0
constructs>=10.0.0
reportlab>=4.0.0
